+++ Assets/scripts/EnemyLogic/Movement/IMovableEnemy.cs
@@ -1,11 +1,10 @@
-﻿using System.Collections.Generic;
-using GridView;
+﻿using GridView;
 using UnityEngine;
 
 namespace EnemyLogic.Movement
 {
     public interface IMovableEnemy
     {
-        List<Vector3> MoveToPlayerDirection(Transform transform, Vector2 targetPosition, Field _field);
+        void MoveToPlayerDirection(Transform transform, Vector2 targetPosition, Field field);
     }
 }
