+++ Assets/scripts/EnemyLogic/Movement/PassiveEnemyMovement.cs
@@ -1,14 +1,13 @@
-﻿using System.Collections.Generic;
-using GridView;
+﻿using GridView;
 using UnityEngine;
 
 namespace EnemyLogic.Movement
 {
     public class PassiveEnemyMovement : IMovableEnemy
     {
-        public List<Vector3> MoveToPlayerDirection(Transform transform, Vector2 targetPosition, Field field)
+        public void MoveToPlayerDirection(Transform transform, Vector2 targetPosition, Field field)
         {
-            return new List<Vector3>();
+            return;
         }
     }
 }
